mainProject/requirements.txt
# django-admin startproject UrbanDjango
# cd DjangoApp
# python manage.py runserver
# python manage.py startapp task1
# python manage.py startapp example2
# python manage.py startapp example3
# python manage.py makemigrations
# python manage.py migrate
# python manage.py shell
django
sqlalchemy
tortoise-orm
fastapi
uvicorn
jinja2
